♻️refactor: UIViewController에 대한 에러 핸들링 extension으로 분리
