+++ Expo1900/Expo1900/Controller/EntryListViewController.swift
@@ -26,19 +26,7 @@ final class EntryListViewController: UIViewController {
         do {
             entryList = try Decoder.decodeJSON(DataAssetNamespace.entryList)
         } catch {
-            let alert = UIAlertController(title: nil, message: nil, preferredStyle: .alert)
-            alert.addAction(UIAlertAction(title: "확인", style: .default))
-            
-            switch error {
-            case DecoderError.notFoundAsset:
-                alert.message = DecoderError.notFoundAsset.description
-            case DecoderError.decodeFailed:
-                alert.message = DecoderError.decodeFailed.description
-            default:
-                alert.message = DecoderError.unexpectedError.description
-            }
-            
-            present(alert, animated: true)
+            showDecoderErrorAlert(error: error)
         }
     }
 }
